[refactor] 불필요한 코드 제거 리팩토링
