+++ src/test/java/domain/CardCalculatorTest.java
@@ -1,9 +1,6 @@
 package domain;
 
-import domain.card.Card;
-import domain.card.CardNumber;
-import domain.card.CardSuitSymbol;
-import domain.card.Cards;
+import domain.card.*;
 import domain.player.Dealer;
 import domain.player.Player;
 import domain.player.User;
